uptaded readme and also added youtube trailer video in the main container
