+++ src/Components/MainContainer.js
@@ -6,14 +6,13 @@ import VideoTitle from "./VideoTitle";
 const MainContainer = () => {
   const movies = useSelector((store) => store.movies?.nowPlayingMovies);
   if (!movies) return;
-  const mainMovie = movies[0];
-const {original_title, overview,id} = mainMovie; 
+  const mainMovie = movies[6];
+  const { original_title, overview, id } = mainMovie;
 
   return (
     <>
-      <div>MainContainer</div>
-      <VideoBackground movie_id={id}/>
       <VideoTitle title={original_title} overview={overview} />
+      <VideoBackground movie_id={id} />
     </>
   );
 };
