+++ src/hooks/useNowPlayingMovies.js
@@ -11,7 +11,6 @@ const useNowPlayingMovies = () => {
         API_OPTIONS
       );
       const json = await data.json();
-      console.log(json.results);
       dispatch(addNowPlayingMovies(json.results))
     };
     useEffect(() => {
